requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-cloud-discoveryengine
python-dotenv
